Generating sentences with 1 atr, 2 atr with AND

+++ Zadanie2_KSR/MembershipFunctions/GaussianFunction.cs
@@ -13,7 +13,7 @@ namespace Zadanie2_KSR.MembershipFunctions
             _width = width;
         }
 
-        public double CountValue(int x)
+        public double CountValue(double x)
         {
             var tmpFraction = (x - _avgX) / _width;
             return Math.Exp(-tmpFraction * tmpFraction);
